+++ src/components/index.tsx
@@ -1,0 +1,4 @@
+export * from "./Footer";
+export * from "./Form";
+export * from "./Header";
+export * from "./Result";
