+++ src/data/index.ts
@@ -1,2 +1,3 @@
 export * from "./2022";
+export * from "./baseTotalReturn";
 export * from "./content"
